@@ -95,50 +95,50 @@
 		"@types/request-stats": "3.0.0",
 		"@types/rimraf": "3.0.0",
 		"@types/seedrandom": "2.4.28",
-		"@types/sharp": "0.26.1",
-		"@types/sinonjs__fake-timers": "6.0.1",
+		"@types/sharp": "0.27.1",
+		"@types/sinonjs__fake-timers": "6.0.2",
 		"@types/speakeasy": "2.0.5",
 		"@types/throttle-debounce": "2.1.0",
 		"@types/tinycolor2": "1.4.2",
 		"@types/tmp": "0.2.0",
 		"@types/uuid": "8.3.0",
 		"@types/web-push": "3.3.0",
-		"@types/webpack": "4.41.25",
+		"@types/webpack": "4.41.26",
 		"@types/webpack-stream": "3.2.11",
 		"@types/websocket": "1.0.1",
 		"@types/ws": "7.4.0",
-		"@typescript-eslint/parser": "4.10.0",
+		"@typescript-eslint/parser": "4.14.2",
 		"@vue/compiler-sfc": "3.0.5",
 		"abort-controller": "3.0.0",
-		"apexcharts": "3.23.1",
+		"apexcharts": "3.24.0",
 		"autobind-decorator": "2.4.0",
 		"autosize": "4.0.2",
 		"autwh": "0.1.0",
-		"aws-sdk": "2.809.0",
+		"aws-sdk": "2.839.0",
 		"bcryptjs": "2.4.3",
 		"blurhash": "1.1.3",
 		"bull": "3.20.0",
 		"cafy": "15.2.1",
-		"cbor": "5.1.0",
+		"cbor": "6.0.1",
 		"chalk": "4.1.0",
 		"chart.js": "2.9.4",
-		"cli-highlight": "2.1.9",
+		"cli-highlight": "2.1.10",
 		"commander": "4.1.1",
 		"content-disposition": "0.5.3",
-		"core-js": "3.8.1",
+		"core-js": "3.8.3",
 		"crc-32": "1.2.0",
 		"css-loader": "5.0.1",
 		"cssnano": "4.1.10",
-		"dateformat": "4.3.1",
+		"dateformat": "4.5.1",
 		"diskusage": "1.1.3",
 		"double-ended-queue": "2.1.0-0",
 		"escape-regexp": "0.0.1",
-		"eslint": "7.17.0",
-		"eslint-plugin-vue": "7.4.1",
+		"eslint": "7.19.0",
+		"eslint-plugin-vue": "7.5.0",
 		"eventemitter3": "4.0.7",
-		"feed": "4.2.1",
+		"feed": "4.2.2",
 		"fibers": "5.0.0",
-		"file-type": "16.1.0",
+		"file-type": "16.2.0",
 		"fluent-ffmpeg": "2.1.2",
 		"glob": "7.1.6",
 		"got": "11.8.1",
@@ -250,22 +247,22 @@
 		"vue": "3.0.5",
 		"vue-color": "2.8.1",
 		"vue-json-pretty": "1.7.1",
-		"vue-loader": "16.0.0",
+		"vue-loader": "16.1.2",
 		"vue-prism-editor": "2.0.0-alpha.2",
-		"vue-router": "4.0.1",
+		"vue-router": "4.0.3",
 		"vue-style-loader": "4.1.2",
 		"vuedraggable": "4.0.1",
 		"web-push": "3.4.4",
-		"webpack": "5.13.0",
-		"webpack-cli": "4.3.1",
+		"webpack": "5.21.0",
+		"webpack-cli": "4.5.0",
 		"websocket": "1.0.33",
-		"ws": "7.4.2",
+		"ws": "7.4.3",
 		"xev": "2.0.1"
 	},
 	"devDependencies": {
 		"@types/chai": "4.2.14",
 		"@types/fluent-ffmpeg": "2.1.16",
-		"chai": "4.2.0",
+		"chai": "4.3.0",
 		"cross-env": "7.0.3"
 	}
 }
