@@ -1,51 +1,55 @@
 <template>
 <div>
-	<MkButton class="kudkigyw" @click="click()" :primary="value.primary">{{ hpml.interpolate(value.text) }}</MkButton>
+	<MkButton class="kudkigyw" @click="click()" :primary="block.primary">{{ hpml.interpolate(block.text) }}</MkButton>
 </div>
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue';
+import { defineComponent, PropType, unref } from 'vue';
 import MkButton from '../ui/button.vue';
 import * as os from '@/os';
+import { ButtonBlock } from '@/scripts/hpml/block';
+import { Hpml } from '@/scripts/hpml/evaluator';
 
 export default defineComponent({
 	components: {
 		MkButton
 	},
 	props: {
-		value: {
+		block: {
+			type: Object as PropType<ButtonBlock>,
 			required: true
 		},
 		hpml: {
+			type: Object as PropType<Hpml>,
 			required: true
 		}
 	},
 	methods: {
 		click() {
-			if (this.value.action === 'dialog') {
+			if (this.block.action === 'dialog') {
 				this.hpml.eval();
 				os.dialog({
-					text: this.hpml.interpolate(this.value.content)
+					text: this.hpml.interpolate(this.block.content)
 				});
-			} else if (this.value.action === 'resetRandom') {
+			} else if (this.block.action === 'resetRandom') {
 				this.hpml.updateRandomSeed(Math.random());
 				this.hpml.eval();
-			} else if (this.value.action === 'pushEvent') {
+			} else if (this.block.action === 'pushEvent') {
 				os.api('page-push', {
 					pageId: this.hpml.page.id,
-					event: this.value.event,
-					...(this.value.var ? {
-						var: this.hpml.vars[this.value.var]
+					event: this.block.event,
+					...(this.block.var ? {
+						var: unref(this.hpml.vars)[this.block.var]
 					} : {})
 				});
 
 				os.dialog({
 					type: 'success',
-					text: this.hpml.interpolate(this.value.message)
+					text: this.hpml.interpolate(this.block.message)
 				});
-			} else if (this.value.action === 'callAiScript') {
-				this.hpml.callAiScript(this.value.fn);
+			} else if (this.block.action === 'callAiScript') {
+				this.hpml.callAiScript(this.block.fn);
 			}
 		}
 	}
