@@ -1,27 +1,29 @@
 <template>
-<div v-show="hpml.vars[value.var]">
-	<XBlock v-for="child in value.children" :value="child" :page="page" :hpml="hpml" :key="child.id" :h="h"/>
+<div v-show="hpml.vars.value[block.var]">
+	<XBlock v-for="child in block.children" :block="child" :hpml="hpml" :key="child.id" :h="h"/>
 </div>
 </template>
 
 <script lang="ts">
-import { defineComponent, defineAsyncComponent } from 'vue';
+import { IfBlock } from '@/scripts/hpml/block';
+import { Hpml } from '@/scripts/hpml/evaluator';
+import { defineComponent, defineAsyncComponent, PropType } from 'vue';
 
 export default defineComponent({
 	components: {
 		XBlock: defineAsyncComponent(() => import('./page.block.vue'))
 	},
 	props: {
-		value: {
+		block: {
+			type: Object as PropType<IfBlock>,
 			required: true
 		},
 		hpml: {
-			required: true
-		},
-		page: {
+			type: Object as PropType<Hpml>,
 			required: true
 		},
 		h: {
+			type: Number,
 			required: true
 		}
 	},
