@@ -5,25 +5,28 @@
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue';
+import { defineComponent, PropType } from 'vue';
 import * as os from '@/os';
+import { ImageBlock } from '@/scripts/hpml/block';
+import { Hpml } from '@/scripts/hpml/evaluator';
 
 export default defineComponent({
 	props: {
-		value: {
+		block: {
+			type: Object as PropType<ImageBlock>,
 			required: true
 		},
-		page: {
+		hpml: {
+			type: Object as PropType<Hpml>,
 			required: true
-		},
+		}
 	},
-	data() {
+	setup(props, ctx) {
+		const image = props.hpml.page.attachedFiles.find(x => x.id === props.block.fileId);
+
 		return {
-			image: null,
+			image
 		};
-	},
-	created() {
-		this.image = this.page.attachedFiles.find(x => x.id === this.value.fileId);
 	}
 });
 </script>
