@@ -6,12 +6,14 @@
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue';
+import { defineComponent, PropType } from 'vue';
 import { faCheck, faPaperPlane } from '@fortawesome/free-solid-svg-icons';
 import MkTextarea from '../ui/textarea.vue';
 import MkButton from '../ui/button.vue';
 import { apiUrl } from '@/config';
 import * as os from '@/os';
+import { PostBlock } from '@/scripts/hpml/block';
+import { Hpml } from '@/scripts/hpml/evaluator';
 
 export default defineComponent({
 	components: {
@@ -19,16 +21,18 @@ export default defineComponent({
 		MkButton,
 	},
 	props: {
-		value: {
+		block: {
+			type: Object as PropType<PostBlock>,
 			required: true
 		},
 		hpml: {
+			type: Object as PropType<Hpml>,
 			required: true
 		}
 	},
 	data() {
 		return {
-			text: this.hpml.interpolate(this.value.text),
+			text: this.hpml.interpolate(this.block.text),
 			posted: false,
 			posting: false,
 			faCheck, faPaperPlane
@@ -37,7 +41,7 @@ export default defineComponent({
 	watch: {
 		'hpml.vars': {
 			handler() {
-				this.text = this.hpml.interpolate(this.value.text);
+				this.text = this.hpml.interpolate(this.block.text);
 			},
 			deep: true
 		}
@@ -45,7 +49,7 @@ export default defineComponent({
 	methods: {
 		upload() {
 			const promise = new Promise((ok) => {
-				const canvas = this.hpml.canvases[this.value.canvasId];
+				const canvas = this.hpml.canvases[this.block.canvasId];
 				canvas.toBlob(blob => {
 					const data = new FormData();
 					data.append('file', blob);
@@ -69,7 +73,7 @@ export default defineComponent({
 		},
 		async post() {
 			this.posting = true;
-			const file = this.value.attachCanvasImage ? await this.upload() : null;
+			const file = this.block.attachCanvasImage ? await this.upload() : null;
 			os.apiWithDialog('notes/create', {
 				text: this.text === '' ? null : this.text,
 				fileIds: file ? [file.id] : undefined,
