@@ -86,8 +86,7 @@ function isPublic(id: string) {
 }
 
 function isFollowers(id: string, actor: IRemoteUser) {
-	return [
-		`${actor.uri}/followers`,
-		// actor.followerUri,	// TODO
-	].includes(id);
+	return (
+		id === (actor.followersUri || `${actor.uri}/followers`)
+	);
 }
