@@ -16,7 +16,7 @@ export default class extends Channel {
 
 			switch (type) {
 				case 'notification': {
-					if (this.muting.includes(body.userId)) return;
+					if (this.muting.has(body.userId)) return;
 					if (body.note && body.note.isHidden) {
 						body.note = await Notes.pack(body.note.id, this.user, {
 							detail: true
@@ -25,7 +25,7 @@ export default class extends Channel {
 					break;
 				}
 				case 'mention': {
-					if (this.muting.includes(body.userId)) return;
+					if (this.muting.has(body.userId)) return;
 					if (body.isHidden) {
 						body = await Notes.pack(body.id, this.user, {
 							detail: true
